requirements.txt
asgiref==3.4.1
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
Pillow==8.3.1
Pygments==2.9.0
pytz==2021.1
sqlparse==0.4.1
psycopg2-binary
gunicorn==20.1.0